--- config/conf.prc
@@ -1,4 +1,0 @@
-win-size 1280 720
-window-title notWiiSports
-show-frame-rate-meter True
-model-cache-dir
