--- imu/utils/constants.py
@@ -1,6 +1,0 @@
-RAD_TO_DEG = 57.29578  # Conversion factor
-G_GAIN = 0.070  # Gyro sensitivity [deg/s/LSB]
-AA = 0.40  # Complementary filter constant to balance gyro-acc influence, modify
-BALL_MASS = 0  # Change to mass constant
-FRICTION = 0  # Change for friction floor
-PIN_MASS = 0  # Change for pin mass
